fix: route student-facing code-block-viewer through /api/run-code
CodeBlockViewer (used in lesson-renderer + presentation modes) was
missed in #931 — it still called emkc.org directly, so students saw
the same "API error: 401" the editor block used to throw.

  - Extract execution logic into a shared `frontend/lib/code-execution.ts`
    so both the editor block and the viewer use the same proxy + Pyodide
    path. Single language config, no more drift.
  - Editor block (code-block.tsx) and viewer (code-block-viewer.tsx) now
    call `executeCode(language, code)` from the shared module.

Also addresses the follow-up suggestion from the previous PR's audit —
the language allowlist now lives in one place.
